postal: start an internal messaging service with postal
The idea is to keep the different API parts independant and to avoid the
model to fiddle with the API. The API can call the model if they want
to, but the model cannot do anything with the API (no access to HTTP, no
access to Socket.IO).

This is currently incomplete as we have to decide where the groups
should live.