displays: handle multiple connections
When a display connect multiple time, the first disconnection was
considered as terminating all instances. This is not true. For example,
if the display detects a problem with the connection before the server,
it will try to open a new connection and the old one will be closed
"later". Or easier to reproduce, when we open a receiver on two
different tabs, closing one should kept the display connected.