receiver: try harder to reconnect on failure
It is not quite clear if a limited number of reconnect attempts mean we
don't reconnect at all after this number of attempts (the source code
seems to reset the logic). Therefore, just bump the number of attempts
to infinity with a maximum delay of 1 minute and reload in case of
catastrophic failure.