chromecast: build the default receiver URL using the right port
While we can't be sure of the receiver URL, we can try to be more
correct by using the port we are listening to. The next step would be to
use our IP. I would have liked to do an UDP connection to some multicast
address on a discarded port for that since it allows us to get the
source IP synchronously without really sending a packet but NodeJS
doesn't expose the `connect()` function.