receiver: properly detect local storage support
When local storage is disabled in the browser, trying to access
`window.localStorage` may trigger a security exception. We take this
case in account for our detection. The detection is not complete because
some other implementations would just silently discard set
data. However, in our case, we just don't care as we are still relying
on other methods even when local storage support is present.