Docs for Chromecast.receiver
Google has hard coded the DNS servers into Chromecasts. 

Therefore if you have Dashkiosk running on an internal only server, and only having internal DNS records, then a Chromecast will not be able to find the server from a domain name. 

Instead you need to use an IP address or intercept the DNS requests for the Chromecast to find the receiver.