docker: don't reinstall all node modules, just link
Using `npm install` in the dist directory was useful if you plan to get
ride of the non-dist stuff. However, we don't do that, so we can just
use a symbolic link. We don't do that because there is no space to get
back because of how Docker use layers.