receiver: remove benchmark and reduce animations by default
The benchmark is silly and doesn't prove anything. We assumed the
benchmark was able to tell if a device was able to smoothly animate
different stuff. This was not the case.

We have now 3 grades:

 - slow (no animation at all)
 - medium (the default, animate spinners but not transitions)
 - fast (animate everything)

Maybe the default could be chosen through the user agent or a
command-line switch. This is not the case. To select a non-default
value, the user has to choose another receiver:

 - /receiver-slow
 - /receiver-medium
 - /receiver-fast